more typing

+++ src/limiter.cpp
@@ -31,7 +31,7 @@
 
 Limiter::Limiter()
 {
-    init( 0.15, 0.50, 0.60 );
+    init( 0.15f, 0.50f, 0.60f );
 }
 
 Limiter::Limiter( float attackMs, float releaseMs, float thresholdDb )
@@ -48,19 +48,19 @@ Limiter::~Limiter()
 
 void Limiter::setAttack( float attackMs )
 {
-    pAttack = ( float ) attackMs;
+    pAttack = attackMs;
     recalculate();
 }
 
 void Limiter::setRelease( float releaseMs )
 {
-    pRelease = ( float ) releaseMs;
+    pRelease = releaseMs;
     recalculate();
 }
 
 void Limiter::setThreshold( float thresholdDb )
 {
-    pTresh = ( float ) thresholdDb;
+    pTresh = thresholdDb;
     recalculate();
 }
 
@@ -73,11 +73,11 @@ float Limiter::getLinearGR()
 
 void Limiter::init( float attackMs, float releaseMs, float thresholdDb )
 {
-    pAttack  = ( float ) attackMs;
-    pRelease = ( float ) releaseMs;
-    pTresh   = ( float ) thresholdDb;
-    pTrim    = ( float ) 0.60;
-    pKnee    = ( float ) 0.40;
+    pAttack  = attackMs;
+    pRelease = releaseMs;
+    pTresh   = thresholdDb;
+    pTrim    = 0.60f;
+    pKnee    = 0.40f;
 
     gain = 1.f;
 
@@ -86,15 +86,15 @@ void Limiter::init( float attackMs, float releaseMs, float thresholdDb )
 
 void Limiter::recalculate()
 {
-    if ( pKnee > 0.5 ) {
+    if ( pKnee > 0.5f ) {
         // soft knee
-        thresh = ( float ) pow( 10.0, 1.f - ( 2.0 * pTresh ));
+        thresh = pow( 10.0f, 1.f - ( 2.f * pTresh ));
     }
     else {
         // hard knee
-        thresh = ( float ) pow( 10.0, ( 2.0 * pTresh ) - 2.0 );
+        thresh = pow( 10.0f, ( 2.f * pTresh ) - 2.0f );
     }
-    trim = ( float )( pow( 10.0, ( 2.0 * pTrim) - 1.f ));
-    att  = ( float )  pow( 10.0, -2.0 * pAttack );
-    rel  = ( float )  pow( 10.0, -2.0 - ( 3.0 * pRelease ));
+    trim = pow( 10.0f, ( 2.f * pTrim ) - 1.f );
+    att  = pow( 10.0f, -2.f * pAttack );
+    rel  = pow( 10.0f, -2.f - ( 3.f * pRelease ));
 }
